feat: Second review comments
New class for custom matchers
