+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/JmxTransTest.java
@@ -37,7 +37,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static io.strimzi.operator.cluster.ResourceUtils.hasEntries;
+import static io.strimzi.operator.cluster.CustomMatchers.hasEntries;
 import static java.util.Collections.singletonMap;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.hasItem;
